\name{genericAllocationProfit}
\docType{methods}
\alias{allocationProfit-methods}
\title{allocationProfit()}
\description{
Methods for generic function \code{allocationProfit}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "climbingTools")}}{
Takes an allocation and coefficient vectors to return a profit value
}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
