\name{findNeighborhood-genericfunction}
\docType{methods}
\alias{findNeighborhood-methods}
\title{Methods for Function \code{findNeighborhood()}}
\description{
Methods for function \code{findNeighborhood(object)} 
}
\section{Methods}{
\describe{

\item{\code{signature(object = "climbingTools")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
