\name{hillclimb-methods}
\docType{methods}
\alias{hillclimb-methods}
\alias{hillclimb,hillclimbing-method}
\title{Methods for Function \code{hillclimb}}
\description{
Generic function associated with method \code{hillclimb(object)}
}
\section{Methods}{
\describe{
\item{\code{hillclimb(object)}}{  method to find mean normalized profits and mean steps to reach the local maxima using a given ascent method, over a given number of landscapes}
}}
