\name{hillclimbboxplot-methods}
\docType{methods}
\alias{hillclimbboxplot-methods}
\alias{hillclimbboxplot,hillclimbing-method}
\title{ Generic Function \code{hillclimbboxplot()}}
\description{
Methods for Generic function \code{hillclimbboxplot()}
}
\section{Methods}{
\describe{

\item{\code{hillclimbboxplot(object))}}{ Method which returns multiple boxplots representing the mean normalized profits for all approaches, for all connections from 1 to 5 }
}}
