\name{hillclimbtable-methods}
\docType{methods}
\alias{hillclimbtable-methods}
\alias{hillclimbtable,hillclimbing-method}
\title{Generic Function \code{hillclimbtable} }
\description{
Methods for generic function \code{hillclimbtable}  ~~
}
\section{Methods}{
\describe{

\item{hillclimbtable}{\code{signature(object = "hillclimbing")}: table containing mean normalized profits for all ascent methods for all connections from one to five. }
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
