\name{landscapePlot-methods}
\docType{methods}
\alias{landscapePlot-methods}
\alias{landscapePlot,hillclimbing-method}
\title{Generic Function \code{landscapePlot}}
\description{
Methods for generic function \code{landscapePlot}
}
\section{Methods}{
\describe{
\item{landscapePlot}{\code{signature(object = "hillclimbing")}: method which inherits the budget from the object to create a randomly generated landscape for three inputs. }
}
}