\name{standardError-methods}
\docType{methods}
\alias{standardError-methods}
\title{ Generic Function \code{standardError}}
\description{
\code{standardError} is a generic function used to calculate the standard error using the method \code{standardError()}
}
\section{Methods}{
\describe{
\item{standardError}{\code{signature(object = "climbingTools")}
{method to calculate standard error}}
}}
