\name{twoD.simplex-methods}
\docType{methods}
\alias{twoD.simplex-methods}
\alias{twoD.simplex,hillclimbing-method}
\title{ Generic Function \code{twoD.simplex}}
\description{
Methods for generic function \code{twoD.simplex} which create a 2D simplex
}
\section{Methods}{
\describe{
\item{twoD.simplex}{\code{signature(object = "hillclimbing")}: method which inherits the budget from the object to create a 2D simplex. }
}
}
